core: move unused-css to computed artifact
