+++ lighthouse-core/test/audits/byte-efficiency/unused-css-rules-test.js
@@ -20,103 +20,6 @@ describe('Best Practices: unused css rules audit', () => {
     return arr.join('');
   }
 
-  describe('#determineContentPreview', () => {
-    function assertLinesContained(actual, expected) {
-      expected.split('\n').forEach(line => {
-        assert.ok(actual.includes(line.trim()), `${line} is found in preview`);
-      });
-    }
-
-    const preview = UnusedCSSAudit.determineContentPreview;
-
-    it('correctly computes short content preview', () => {
-      const shortContent = `
-        html, body {
-          background: green;
-        }
-      `.trim();
-
-      assertLinesContained(preview(shortContent), shortContent);
-    });
-
-    it('correctly computes long content preview', () => {
-      const longContent = `
-        body {
-          color: white;
-        }
-
-        html {
-          content: '${generate('random', 50)}';
-        }
-      `.trim();
-
-      assertLinesContained(preview(longContent), `
-        body {
-          color: white;
-        } ...
-      `.trim());
-    });
-
-    it('correctly computes long rule content preview', () => {
-      const longContent = `
-        body {
-          color: white;
-          font-size: 20px;
-          content: '${generate('random', 50)}';
-        }
-      `.trim();
-
-      assertLinesContained(preview(longContent), `
-        body {
-          color: white;
-          font-size: 20px; ... } ...
-      `.trim());
-    });
-
-    it('correctly computes long comment content preview', () => {
-      const longContent = `
-      /**
-       * @license ${generate('a', 100)}
-       */
-      `.trim();
-
-      assert.ok(/aaa\.\.\./.test(preview(longContent)));
-    });
-  });
-
-  describe('#mapSheetToResult', () => {
-    let baseSheet;
-    const baseUrl = 'http://g.co/';
-
-    function map(overrides, url = baseUrl) {
-      if (overrides.header && overrides.header.sourceURL) {
-        overrides.header.sourceURL = baseUrl + overrides.header.sourceURL;
-      }
-      return UnusedCSSAudit.mapSheetToResult(Object.assign(baseSheet, overrides), url);
-    }
-
-    beforeEach(() => {
-      baseSheet = {
-        header: {sourceURL: baseUrl},
-        content: 'dummy',
-        usedRules: [],
-      };
-    });
-
-    it('correctly computes wastedBytes', () => {
-      assert.equal(map({usedRules: []}).wastedPercent, 100);
-      assert.equal(map({usedRules: [{startOffset: 0, endOffset: 3}]}).wastedPercent, 40);
-      assert.equal(map({usedRules: [{startOffset: 0, endOffset: 5}]}).wastedPercent, 0);
-    });
-
-    it('correctly computes url', () => {
-      const expectedPreview = 'dummy';
-      assert.strictEqual(map({header: {sourceURL: ''}}).url, expectedPreview);
-      assert.strictEqual(map({header: {sourceURL: 'a'}}, 'http://g.co/a').url, expectedPreview);
-      assert.equal(map({header: {sourceURL: 'foobar'}}).url, 'http://g.co/foobar');
-    });
-  });
-
   describe('#audit', () => {
     const networkRecords = [
       {
@@ -127,6 +30,8 @@ describe('Best Practices: unused css rules audit', () => {
       },
     ];
 
+    const context = {computedCache: new Map()};
+
     function getArtifacts({CSSUsage}) {
       return {
         devtoolsLogs: {defaultPass: networkRecordsToDevtoolsLog(networkRecords)},
@@ -135,10 +40,14 @@ describe('Best Practices: unused css rules audit', () => {
       };
     }
 
+    beforeEach(() => {
+      context.computedCache.clear();
+    });
+
     it('ignores missing stylesheets', () => {
       return UnusedCSSAudit.audit_(getArtifacts({
         CSSUsage: {rules: [{styleSheetId: 'a', used: false}], stylesheets: []},
-      }), networkRecords).then(result => {
+      }), networkRecords, context).then(result => {
         assert.equal(result.items.length, 0);
       });
     });
@@ -159,7 +68,7 @@ describe('Best Practices: unused css rules audit', () => {
             content: '.my.favorite.selector { rule: content; }',
           },
         ]},
-      }), networkRecords).then(result => {
+      }), networkRecords, context).then(result => {
         assert.equal(result.items.length, 0);
       });
     });
@@ -183,7 +92,7 @@ describe('Best Practices: unused css rules audit', () => {
             content: `${generate('123', 450)}`, // will be filtered out
           },
         ]},
-      }), networkRecords).then(result => {
+      }), networkRecords, context).then(result => {
         assert.equal(result.items.length, 2);
         assert.equal(result.items[0].totalBytes, 100 * 1024);
         assert.equal(result.items[1].totalBytes, 40000 * 3 * 0.2);
@@ -215,7 +124,7 @@ describe('Best Practices: unused css rules audit', () => {
           resourceSize: 0,
           resourceType: 'Document',
         },
-      ]);
+      ], context);
 
       expect(result.items).toMatchObject([
         {totalBytes: 40000 * 3 * 0.5, wastedPercent: 50},
@@ -249,7 +158,7 @@ describe('Best Practices: unused css rules audit', () => {
             content: '       ',
           },
         ]},
-      }), networkRecords).then(result => {
+      }), networkRecords, context).then(result => {
         assert.equal(result.items.length, 1);
         assert.equal(Math.floor(result.items[0].wastedPercent), 33);
       });
